+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `nxos_vpc_interface` resource and data source
 - Add `nxos_port_channel_interface` resource and data source
 - Add `nxos_port_channel_interface_vrf` resource and data source
+- BREAKING CHANGE: Remove `ctrl` attribute of `nxos_spanning_tree_interface` resource and data source
 
 ## 0.3.23